refactor: add typescript to controller and appLoader components
