+++ src/components/controller/interfaces.ts
@@ -37,4 +37,5 @@ export type CallbackType<T> = (data?: T) => void;
 
 export type Options = {
   apiKey?: string;
+  sources?: string;
 };
