--- src/index.js
@@ -1,5 +1,0 @@
-import App from './components/app/app';
-import './global.css';
-
-const app = new App();
-app.start();
